Move addon filters into their own subpackages
Addon filters moved into their own subpackages. This keeps things clean
and prevents a "split package" situation with OSGi deployments.
