+++ scalate-pegdown/src/main/scala/org/fusesource/scalate/filter/pegdown/PegDownFilter.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fusesource.scalate.filter
+package org.fusesource.scalate.filter.pegdown
 
 import org.fusesource.scalate.{TemplateEngineAddOn, RenderContext, TemplateEngine}
 import org.fusesource.scalate.filter.Filter
